requirements.txt
streamlit
opencv-python-headless
numpy
pandas
pillow
matplotlib